+++ contracts/contracts/Operations.sol
@@ -50,12 +50,13 @@ library Operations {
         address owner;
     }
 
-    uint public constant PACKED_DEPOSIT_PUBDATA_BYTES = ACCOUNT_ID_BYTES + TOKEN_BYTES + AMOUNT_BYTES + ADDRESS_BYTES;
+    uint256 public constant PACKED_DEPOSIT_PUBDATA_BYTES =
+        ACCOUNT_ID_BYTES + TOKEN_BYTES + AMOUNT_BYTES + ADDRESS_BYTES;
 
     /// Deserialize deposit pubdata
     function readDepositPubdata(bytes memory _data) internal pure returns (Deposit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = 0;
+        uint256 offset = 0;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
@@ -91,11 +92,12 @@ library Operations {
         uint128 amount;
     }
 
-    uint public constant PACKED_FULL_EXIT_PUBDATA_BYTES = ACCOUNT_ID_BYTES + ADDRESS_BYTES + TOKEN_BYTES + AMOUNT_BYTES;
+    uint256 public constant PACKED_FULL_EXIT_PUBDATA_BYTES =
+        ACCOUNT_ID_BYTES + ADDRESS_BYTES + TOKEN_BYTES + AMOUNT_BYTES;
 
     function readFullExitPubdata(bytes memory _data) internal pure returns (FullExit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = 0;
+        uint256 offset = 0;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
         (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
@@ -116,8 +118,8 @@ library Operations {
     /// @notice Check that full exit pubdata from request and block matches
     function fullExitPubdataMatch(bytes memory _lhs, bytes memory _rhs) internal pure returns (bool) {
         // `amount` is ignored because it is present in block pubdata but not in priority queue
-        uint lhs = Bytes.trim(_lhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
-        uint rhs = Bytes.trim(_rhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
+        uint256 lhs = Bytes.trim(_lhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
+        uint256 rhs = Bytes.trim(_rhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
         return lhs == rhs;
     }
 
@@ -131,13 +133,13 @@ library Operations {
         address owner;
     }
 
-    function readPartialExitPubdata(bytes memory _data, uint _offset)
+    function readPartialExitPubdata(bytes memory _data, uint256 _offset)
         internal
         pure
         returns (PartialExit memory parsed)
     {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = _offset + ACCOUNT_ID_BYTES; // accountId (ignored)
+        uint256 offset = _offset + ACCOUNT_ID_BYTES; // accountId (ignored)
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
         offset += FEE_BYTES; // fee (ignored)
@@ -165,9 +167,13 @@ library Operations {
         address target;
     }
 
-    function readForcedExitPubdata(bytes memory _data, uint _offset) internal pure returns (ForcedExit memory parsed) {
+    function readForcedExitPubdata(bytes memory _data, uint256 _offset)
+        internal
+        pure
+        returns (ForcedExit memory parsed)
+    {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = _offset + ACCOUNT_ID_BYTES * 2; // initiatorAccountId + targetAccountId (ignored)
+        uint256 offset = _offset + ACCOUNT_ID_BYTES * 2; // initiatorAccountId + targetAccountId (ignored)
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
         offset += FEE_BYTES; // fee (ignored)
@@ -196,12 +202,12 @@ library Operations {
         //uint16 fee; -- present in pubdata, ignored at serialization
     }
 
-    function readChangePubKeyPubdata(bytes memory _data, uint _offset)
+    function readChangePubKeyPubdata(bytes memory _data, uint256 _offset)
         internal
         pure
         returns (ChangePubKey memory parsed)
     {
-        uint offset = _offset;
+        uint256 offset = _offset;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
         (offset, parsed.pubKeyHash) = Bytes.readBytes20(_data, offset); // pubKeyHash
         (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
@@ -210,7 +216,7 @@ library Operations {
 
     // Withdrawal data process
 
-    function readWithdrawalData(bytes memory _data, uint _offset)
+    function readWithdrawalData(bytes memory _data, uint256 _offset)
         internal
         pure
         returns (
@@ -220,7 +226,7 @@ library Operations {
             uint128 _amount
         )
     {
-        uint offset = _offset;
+        uint256 offset = _offset;
         (offset, _addToPendingWithdrawalsQueue) = Bytes.readBool(_data, offset);
         (offset, _to) = Bytes.readAddress(_data, offset);
         (offset, _tokenId) = Bytes.readUInt16(_data, offset);
