+++ contracts/dev-contracts/tokens/MintableERC20NoTransferReturnValueTest.sol
@@ -29,29 +29,29 @@ import "./SafeMath.sol";
  * allowances. See {IERC20-approve}.
  */
 contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20NoTransferReturnValueTest {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
-    mapping(address => uint) private _balances;
+    mapping(address => uint256) private _balances;
 
-    mapping(address => mapping(address => uint)) private _allowances;
+    mapping(address => mapping(address => uint256)) private _allowances;
 
-    uint private _totalSupply;
+    uint256 private _totalSupply;
 
-    function mint(address to, uint amount) external {
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
 
     /**
      * @dev See {IERC20-totalSupply}.
      */
-    function totalSupply() public view returns (uint) {
+    function totalSupply() public view returns (uint256) {
         return _totalSupply;
     }
 
     /**
      * @dev See {IERC20-balanceOf}.
      */
-    function balanceOf(address account) public view returns (uint) {
+    function balanceOf(address account) public view returns (uint256) {
         return _balances[account];
     }
 
@@ -63,14 +63,14 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      * - `recipient` cannot be the zero address.
      * - the caller must have a balance of at least `amount`.
      */
-    function transfer(address recipient, uint amount) public {
+    function transfer(address recipient, uint256 amount) public {
         _transfer(_msgSender(), recipient, amount);
     }
 
     /**
      * @dev See {IERC20-allowance}.
      */
-    function allowance(address owner, address spender) public view returns (uint) {
+    function allowance(address owner, address spender) public view returns (uint256) {
         return _allowances[owner][spender];
     }
 
@@ -81,7 +81,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      *
      * - `spender` cannot be the zero address.
      */
-    function approve(address spender, uint amount) public returns (bool) {
+    function approve(address spender, uint256 amount) public returns (bool) {
         _approve(_msgSender(), spender, amount);
         return true;
     }
@@ -101,7 +101,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
     function transferFrom(
         address sender,
         address recipient,
-        uint amount
+        uint256 amount
     ) public {
         _transfer(sender, recipient, amount);
         _approve(
@@ -123,7 +123,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      *
      * - `spender` cannot be the zero address.
      */
-    function increaseAllowance(address spender, uint addedValue) public returns (bool) {
+    function increaseAllowance(address spender, uint256 addedValue) public returns (bool) {
         _approve(_msgSender(), spender, _allowances[_msgSender()][spender].add(addedValue));
         return true;
     }
@@ -142,7 +142,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      * - `spender` must have allowance for the caller of at least
      * `subtractedValue`.
      */
-    function decreaseAllowance(address spender, uint subtractedValue) public returns (bool) {
+    function decreaseAllowance(address spender, uint256 subtractedValue) public returns (bool) {
         _approve(
             _msgSender(),
             spender,
@@ -168,7 +168,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
     function _transfer(
         address sender,
         address recipient,
-        uint amount
+        uint256 amount
     ) internal {
         require(sender != address(0), "ERC20: transfer from the zero address");
         require(recipient != address(0), "ERC20: transfer to the zero address");
@@ -187,7 +187,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      *
      * - `to` cannot be the zero address.
      */
-    function _mint(address account, uint amount) internal {
+    function _mint(address account, uint256 amount) internal {
         require(account != address(0), "ERC20: mint to the zero address");
 
         _totalSupply = _totalSupply.add(amount);
@@ -206,7 +206,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      * - `account` cannot be the zero address.
      * - `account` must have at least `amount` tokens.
      */
-    function _burn(address account, uint amount) internal {
+    function _burn(address account, uint256 amount) internal {
         require(account != address(0), "ERC20: burn from the zero address");
 
         _balances[account] = _balances[account].sub(amount, "ERC20: burn amount exceeds balance");
@@ -230,7 +230,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
     function _approve(
         address owner,
         address spender,
-        uint amount
+        uint256 amount
     ) internal {
         require(owner != address(0), "ERC20: approve from the zero address");
         require(spender != address(0), "ERC20: approve to the zero address");
@@ -245,7 +245,7 @@ contract MintableERC20NoTransferReturnValueTest is ContextTest, MintableIERC20No
      *
      * See {_burn} and {_approve}.
      */
-    function _burnFrom(address account, uint amount) internal {
+    function _burnFrom(address account, uint256 amount) internal {
         _burn(account, amount);
         _approve(
             account,
